requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
django-mptt==0.8.3
django-pipeline==1.6.8
django-reversion==1.10.2
Django==1.8.12            # via django-mptt, django-reversion
lxml==3.6.0
psycopg2==2.6.1
pytz==2016.3
